# Conflicts:
#	docs/project/Plugins.md
#	src/ATE_sammy/ate_sammy/migration/migration_scripts/migration_version_8.py
#	src/ATE_sammy/ate_sammy/templates/test_base_template.jinja2
#	src/ATE_sammy/ate_sammy/verbs/generate.py
#	src/ATE_spyder/ate_spyder/widgets/navigation.py
#	src/Plugins/semi_ate_testers/semi_ate_testers/Flatcache/Flatcache.py

+++ src/ATE_sammy/ate_sammy/migration/migration_scripts/migration_version_8.py
@@ -71,5 +71,9 @@ class MigrationVersion8(MigratorBase):
         for key, values in target_info.items():
             self.write_configuration(key, values)
 
-        import os
-        os.remove(file_name)
+        file_base_name = Path(file_name).stem
+        parent_folder_name = Path(file_name).parent.name
+
+        if (file_base_name == parent_folder_name):
+            import os
+            os.remove(file_name)
